Revamp quickstart
Leveraging https://github.com/readysettech/readyset-public-docs/pull/12,
this PR rewrites the quickstart as a sequence of step and uses a sample
data set to help demonstrate some basic ReadySet features.

This PR also tweaks other parts of the site so that we can
publish it docs.readyset.io and work from this repo going forward.