Rewrite quickstart
This new verison uses Docker to start Postgres and Readyset,
loads a larger dataset from the IMDB sample, and uses
a Python app to run queries against both the database and
ReadySet and print latencies for comparison.

The hope is that this new, more manual approach is still
easy but lets us expose more about how you run and
interact with ReadySet.

For now, we're focusing just on Postgres. Once this lands,
we can work on a MySQL version.

This PR also removes references to the deprecated Grafana
dashboard.

Fixes DOC-65