Revise the quickstart to use Docker Compose
Now that we have manual deployment guidance and CLI references
docs, it's no longer critical for us to education users on
ReadySet options in the quickstart. So this PR refactors
the quickstart to use Docker Compose to:

- Start Postgres
- Load sample data
- Start ReadySet
- Create a container for the Python app

The hope is that this gets users querying and caching quicker.

This PR also:

- Removes MySQL guidance. We can work to add that back if we see
  users wanting it.
- Removes guidance for running with an existing local db. We can
  work to add that back if we see users wanting it.
- Significantly reduces the size of the sample dataset
  to ensure that the snapshotting process doesn't cause users to
  give up.
- Retires the text version of the tutorial in favor of a fully
  in-browser "Playground".

Once this merges, there are some follow-ups:
- Look into updating the Instruqt "playground" as necessary.
- Update the readyset README to correctly reference the new
  quickstart.

Fixes: DOC-116