Revert the DEPLOYMENT_ENV value
We changed telemetry to accept random DEPLOYMENT_ENV values,
but that change isn't in the Jan release, which the quickstart
currently uses. This PR temporarily reverts to the previous
hard-coded `quickstart_docker` value. Once the Feb release
is out, we'll update the quickstart to pull that image and
then will have the flexibility to change the DEPLOYMENT_ENV
value.