docs: Add new quickstart compose file
Add a compose file to be used with the new quickstart, to run ReadySet
either by itself (with an external upstream and application), or
alongside a user's existing compose project.

It takes the following arguments in the form of environment variables:
RS_UPSTREAM_DB_URL: required
RS_PORT: optional, default=3307
RS_GRAFANA_PORT: optional, default=4000

Also, add Dockerfiles for Grafana and Prometheus that bundle relevant
config files, so that those don't need to be manually distributed to
users and bind-mounted into the containers at run time. These
Dockerfiles are used to build images to be published to our public ECR.

Change-Id: I4506b3089b62dd6706738b44665d95d309e3505b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4877
Tested-by: Buildkite CI
Reviewed-by: Ian Meyer <ian@readyset.io>