docs: Add mysql and postgres quickstarts without sample data
Rename the existing mysql and postgres database-included quickstarts
to compose.*.imdb.yml, to denote that they are to be bundled with the
corresponding imdb-*.sql sample data.

Add another two quickstart compose files that are identical to the
above, except that they do not mount the sample data files. This is to
support the quickstart docs giving the user the choice between:
- deploying readyset + grafana
- deploying readyset, grafana, and [mysql | postgres] with IMDB sample
  data
- deploying readyset, grafana, and [mysql | postgres] without IMDB
  sample data

Change-Id: I75492fc415c4c7f8b4b53c252fe7e0bb8a474722
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4968
Tested-by: Buildkite CI
Reviewed-by: David Kim <david@readyset.io>