quickstart: Explicitly add host.docker.internal host mapping
Expose the Docker host as host.docker.internal to the readyset
container. This allows the user to connect to an upstream database
running on the Docker host by providing that hostname in the URL,
without having to manually query the Docker gateway IP.

This hostname mapping is provided by default when using Docker Desktop,
but not by plain Docker Engine; we make this mapping explicit to provide
it in the latter case.

Tested on:
- Docker 24.0.2 / Compose 2.18.1 / Ubuntu 22.04
- Docker Desktop for Mac 4.19.0 / Docker 23.0.5 / Compose 2.17.3

Change-Id: I070bef03983a8f19ca976ea4c0cdc86c9625590c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5017
Tested-by: Buildkite CI
Reviewed-by: Ian Meyer <ian@readyset.io>