server: Rename quorum flag to min-workers
The word "quorum" is a little overloaded, especially in the distsys
space, and it's unclear what it actually means in the context of
ReadySet - where it's the number of healthy workers we'll wait to join
the cluster before we start trying to run domains. To clarify things a
bit, this commit renames the flag to `min_workers` (updating everything
that references it to use the new name).

Change-Id: Ib2154529fe8ea7a979c6003c6306e2d489133899
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5409
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>